doxygen: Tailor configuration file for libgpio
Tailor the doxygen configuration file so it works for the libgpio
repository.  This includes generating HTML output only and putting
that output in the 'docs' directory.

Note that normally 'OUTPUT_DIRECTORY' would be set to 'docs' and
'HTML_OUTPUT' would be left as 'html'.  This would put the HTML
output in 'docs/html'.  Unfortunately, github pages require that
the HTML be in the 'docs' directory and not in a subdirectory
(i.e., in 'docs' itself and not in 'docs/html').  So, to keep all
of the documentation-related files in one place, the configuration
file and the HTML output will be kept in the 'docs' directory even
though its a bit messy.  This is accomplished by leaving
'OUTPUT_DIRECTORY' blank and setting 'HTML_OUTPUT' to "docs".

Signed-off-by: Mark Greer <mgreer@animalcreek.com>