doxygen: Add a main doxygen page
Add a main page for doxygen to help users understand the
purpose of the code in this repository.  This file will
also be put in the 'docs' directory along with the doxygen
configuration file and the HTML that it generates.  This
is messy but keeps all of the documentation-related files
under 'docs'.

Note that github pages requires that the HTML be in the
'docs' directory and not in a subdirectory like 'docs/html'.

Signed-off-by: Mark Greer <mgreer@animalcreek.com>