examples: irq_mon: Add example program to monitor GPIO IRQs
Add an example program that simply waits for a GPIO interrupts
to be generated and prints out a message each time.  This can
serve as a useful example on how to use 'libgpio' and how to
handle GPIO interrupts from userspace (using select()).

Note that dependency files will now be made (i.e., .d files) so
they were added to .gitignore.

Signed-off-by: Mark Greer <mgreer@animalcreek.com>